+++ tests/test_ratelimiter.py
@@ -1,10 +1,16 @@
 # import pytest
 
-from gemeaux import ConnectionLimiter, NoRateLimiter, RateLimiter, SpeedLimiter, SpeedAndConnectionLimiter
+from gemeaux import (
+    ConnectionLimiter,
+    NoRateLimiter,
+    RateLimiter,
+    SpeedAndConnectionLimiter,
+    SpeedLimiter,
+)
 
 
 def test_connection_rate_limiter():
-    for rl in [ ConnectionLimiter(), SpeedAndConnectionLimiter()]:
+    for rl in [ConnectionLimiter(), SpeedAndConnectionLimiter()]:
         for i in range(0, 9):
             assert rl.AddNewConnection("::1") is True
 
@@ -13,7 +19,7 @@ def test_connection_rate_limiter():
 
 
 def test_connection_rate_limiter_with_reset():
-    for rl in [ ConnectionLimiter(), SpeedAndConnectionLimiter()]:
+    for rl in [ConnectionLimiter(), SpeedAndConnectionLimiter()]:
         for i in range(0, 9):
             assert rl.AddNewConnection("::1") is True
         rl.ResetClientList()
@@ -44,9 +50,15 @@ def test_speed_limiter_with_reset():
             assert rl.GetToken("::1", MAX_DOWNLOAD_LIMIT_PER_MINUTE / 10) is True
         rl.ResetClientList()
         if isinstance(rl, SpeedAndConnectionLimiter):
-            assert abs(rl.sl.tokenDict["::1"] * 4 - MAX_DOWNLOAD_LIMIT_PER_MINUTE / 10 * 8) < 1
+            assert (
+                abs(rl.sl.tokenDict["::1"] * 4 - MAX_DOWNLOAD_LIMIT_PER_MINUTE / 10 * 8)
+                < 1
+            )
         else:
-            assert abs(rl.tokenDict["::1"] * 4 - MAX_DOWNLOAD_LIMIT_PER_MINUTE / 10 * 8) < 1
+            assert (
+                abs(rl.tokenDict["::1"] * 4 - MAX_DOWNLOAD_LIMIT_PER_MINUTE / 10 * 8)
+                < 1
+            )
         for i in range(0, 4):
             rl.ResetClientList()
         if isinstance(rl, SpeedAndConnectionLimiter):
@@ -64,3 +76,59 @@ def test_no_rate_limiter():
 
     for i in range(0, 10):
         assert rl.GetToken("::1", 1) is True
+
+
+def test_violation_checker():
+    for limiter in [RateLimiter(), NoRateLimiter()]:
+        for i in range(0, 11):
+            assert limiter.AddNewConnection("::1", 1) is True
+            assert limiter.IsClientInViolation("::1") is False
+
+
+def test_violation_limiter():
+    for limiter in [ConnectionLimiter(), SpeedLimiter(), SpeedAndConnectionLimiter()]:
+        for i in range(0, 9):
+            assert limiter.IsClientInViolation("::1") is False
+
+
+def test_violation_limiter_true():
+    for limiter in [ConnectionLimiter(), SpeedAndConnectionLimiter()]:
+        for i in range(0, 9):
+            assert limiter.AddNewConnection("::1") is True
+
+        assert limiter.AddNewConnection("::1") is False
+        assert limiter.IsClientInViolation("::1") is True
+
+
+def test_get_penalty_time_false():
+    for limiter in [
+        RateLimiter(),
+        NoRateLimiter(),
+        ConnectionLimiter(),
+        SpeedLimiter(),
+        SpeedAndConnectionLimiter(),
+    ]:
+        assert limiter.GetPenaltyTime("::1") == 0
+
+
+def test_get_penalty_time_for_connection_true():
+    for limiter in [ConnectionLimiter(), SpeedAndConnectionLimiter()]:
+        for i in range(0, 9):
+            assert limiter.AddNewConnection("::1") is True
+
+        assert limiter.AddNewConnection("::1") is False
+        assert limiter.IsClientInViolation("::1") is True
+        assert limiter.GetPenaltyTime("::1") > 0
+
+
+def test_get_penalty_time_for_speed_true():
+    MAX_DOWNLOAD_LIMIT_PER_MINUTE = int(
+        SpeedLimiter().MAX_DOWNLOAD_LIMIT_PER_MINUTE / 10 + 0.5
+    )
+    for limiter in [SpeedLimiter(), SpeedAndConnectionLimiter()]:
+        for i in range(0, 9):
+            assert limiter.GetToken("::1", MAX_DOWNLOAD_LIMIT_PER_MINUTE) is True
+
+        assert limiter.GetToken("::1", MAX_DOWNLOAD_LIMIT_PER_MINUTE) is False
+        assert limiter.IsClientInViolation("::1") is True
+        assert limiter.GetPenaltyTime("::1") > 0